update notes import and export file type to jsonl
Changes the file type of notes exports to .jsonl (JSON Lines), which
is a more common format used by compatible wallets as specified in
BIP-329.

For example, exporting labels with Sparrow Wallet creates a .jsonl,
which would previously have to be renamed manually in order to allow
importing it in the BitBoxApp.

The supported import file types are updated accordingly, still allowing
.txt imports for exports created prior to this change.