frontend/ios: enable swipe navigation gestures
Enables navigating the app by swiping from the display edge, similar
to the system back button on Android. This makes navigating the app
especially easier on larger iPhones, as the swipe to go back is much
faster and more natural than reaching for the back button in the top
left corner.

Note that this also enables navigation where there is currently no
dedicated back button. For example, navigating from the portfolio tab
to the accounts tab will allow you to "swipe back" to the portfolio tab.

Because this uses the native iOS navigation gestures, and there is no
option to only enable backward navigation, it is also possible to swipe
from the right display edge to navigate "forward".